requirements.txt
annotated-types==0.5.0
boto3==1.28.35
botocore==1.31.35
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
jmespath==1.0.1
numpy==1.25.2
pandas==2.0.3
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16